+++ Assets/VRMaintenanceResearch/Editor/ResearchFoundationEditorTests.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Globalization;
 using System.IO;
+using System.Threading;
 using NUnit.Framework;
 using UnityEngine;
 
@@ -64,6 +66,62 @@ namespace TMUVR.MaintenanceResearch
             Assert.That(logger.IsStarted, Is.True);
             logger.EndSession("SecondComplete");
 
+            UnityEngine.Object.DestroyImmediate(sessionObject);
+        }
+
+        [Test]
+        public void TaskEventTimestampIsRelativeToTaskStart()
+        {
+            var sessionObject = new GameObject("TaskTimestampTest");
+            var logger = sessionObject.AddComponent<ResearchLogService>();
+            var definition = ScriptableObject.CreateInstance<ResearchTaskDefinition>();
+            definition.taskId = ResearchTaskId.Computer;
+            definition.layoutId = "timestamp-layout";
+            definition.taskContext = "timestamp-test";
+            var config = new ResearchSessionConfig { participantCode = "TEST_TIME", sessionId = "task_time_" + Guid.NewGuid().ToString("N").Substring(0, 8), developmentMode = true };
+
+            Assert.That(logger.BeginSession(config, out var error), Is.True, error);
+            Thread.Sleep(25);
+            logger.BeginTask(definition, 1);
+            logger.EndTask(ResearchTaskId.Computer, TaskState.Completed);
+            logger.EndSession("TestCompleted");
+
+            var firstEvent = File.ReadAllLines(Path.Combine(logger.CurrentDataFolder, "Computer", "events.csv"))[1].Split(',');
+            Assert.That(double.Parse(firstEvent[11], CultureInfo.InvariantCulture), Is.LessThan(0.1d));
+
+            UnityEngine.Object.DestroyImmediate(definition);
+            UnityEngine.Object.DestroyImmediate(sessionObject);
+        }
+
+        [Test]
+        public void SessionSummaryContainsOneRowPerCompletedTask()
+        {
+            var sessionObject = new GameObject("SummaryRowsTest");
+            var logger = sessionObject.AddComponent<ResearchLogService>();
+            var computer = ScriptableObject.CreateInstance<ResearchTaskDefinition>();
+            var fan = ScriptableObject.CreateInstance<ResearchTaskDefinition>();
+            computer.taskId = ResearchTaskId.Computer;
+            computer.taskContext = "computer";
+            computer.layoutId = "computer-layout";
+            fan.taskId = ResearchTaskId.Fan;
+            fan.taskContext = "fan";
+            fan.layoutId = "fan-layout";
+            var config = new ResearchSessionConfig { participantCode = "TEST_SUMMARY", sessionId = "summary_rows_" + Guid.NewGuid().ToString("N").Substring(0, 8), developmentMode = true };
+
+            Assert.That(logger.BeginSession(config, out var error), Is.True, error);
+            logger.BeginTask(computer, 1);
+            logger.EndTask(ResearchTaskId.Computer, TaskState.Completed);
+            logger.BeginTask(fan, 1);
+            logger.EndTask(ResearchTaskId.Fan, TaskState.Completed);
+            logger.EndSession("TestCompleted");
+
+            var rows = File.ReadAllLines(Path.Combine(logger.CurrentDataFolder, "task_summary.csv"));
+            Assert.That(rows, Has.Length.EqualTo(3));
+            Assert.That(rows[1], Does.Contain(",Computer,"));
+            Assert.That(rows[2], Does.Contain(",Fan,"));
+
+            UnityEngine.Object.DestroyImmediate(computer);
+            UnityEngine.Object.DestroyImmediate(fan);
             UnityEngine.Object.DestroyImmediate(sessionObject);
         }
     }